created by userid from long to int

+++ Database/dbo/Tables/UM_Bureau.sql
@@ -4,11 +4,13 @@
     [Name]            NVARCHAR (200) NULL,
     [Description]     NVARCHAR (MAX) NULL,
     [IsActive]        BIT            NOT NULL,
-    [CreatedByUserId] BIGINT         NOT NULL,
+    [CreatedByUserId] INT            NOT NULL,
     [CreatedDate]     DATETIME       NOT NULL,
     [DivisionId]      INT            NULL,
     [GroupId]         INT            NULL,
     CONSTRAINT [PK_UM_Bureau] PRIMARY KEY CLUSTERED ([BureauId] ASC),
     CONSTRAINT [FK_UM_Bureau_UM_Division] FOREIGN KEY ([DivisionId]) REFERENCES [dbo].[UM_Division] ([DivisionId])
 );
+
+
