+++ Database/dbo/Tables/UM_Division.sql
@@ -4,8 +4,10 @@
     [Name]            NVARCHAR (200) NULL,
     [Description]     NVARCHAR (MAX) NULL,
     [IsActive]        BIT            NOT NULL,
-    [CreatedByUserId] BIGINT         NOT NULL,
+    [CreatedByUserId] INT            NOT NULL,
     [CreatedDate]     DATETIME       NOT NULL,
     CONSTRAINT [PK_UM_Division] PRIMARY KEY CLUSTERED ([DivisionId] ASC)
 );
+
+
