+++ Database/dbo/Tables/UM_Section.sql
@@ -4,10 +4,12 @@
     [Name]            NVARCHAR (200) NULL,
     [Description]     NVARCHAR (MAX) NULL,
     [IsActive]        BIT            NOT NULL,
-    [CreatedByUserId] BIGINT         NOT NULL,
+    [CreatedByUserId] INT            NOT NULL,
     [CreatedDate]     DATETIME       NOT NULL,
     [DepartmentId]    INT            NULL,
     CONSTRAINT [PK_UM_Section] PRIMARY KEY CLUSTERED ([SectionId] ASC),
     CONSTRAINT [FK_UM_Section_UM_Department] FOREIGN KEY ([DepartmentId]) REFERENCES [dbo].[UM_Department] ([DepartmentId])
 );
+
+
